add watching manager hook
