+++ src/context/app.tsx
@@ -1,14 +1,14 @@
 'use client';
 
-import { ReactNode, useCallback, useEffect, useState } from 'react';
-import { useSearchParams } from 'next/navigation';
-import { Category, Stream, User } from '@/types/twitch';
+import { ReactNode, useEffect, useState } from 'react';
+import { Category, User } from '@/types/twitch';
 import { PlayerView } from '@/types/state';
 import { createContext } from 'react';
 import {
   getLsFollowedCategories,
   setLsFollowedCategories,
 } from '@/lib/local-storage';
+import { useWatchingManager } from '@/hooks/watching-manager';
 
 interface AppContextType {
   accessToken: string | undefined;
@@ -37,8 +37,6 @@ export const AppContext = createContext<AppContextType | undefined>(undefined);
 
 export function AppProvider({ children }: { children: ReactNode }) {
   // global helpers
-  const searchParams = useSearchParams();
-
   const initialWatching: string[] = [];
   const initialChat = '';
   const initialView = PlayerView.Grid;
@@ -53,68 +51,14 @@ export function AppProvider({ children }: { children: ReactNode }) {
   const [followedCategories, setFollowedCategories] = useState<Category[]>([]);
   const [updatePath, setUpdatePath] = useState<boolean>(false);
 
-  // update active chat if removed
-  useEffect(() => {
-    let found = watching.find((e) => e == activeChat);
-    if (!found) {
-      setActiveChat(watching[0]);
-    }
-  }, [watching, activeChat]);
-
-  // add channel to watching state
-  function addWatching(channel: string) {
-    if (watching.includes(channel)) {
-      // add highlight animation
-      let channelDiv = document.getElementById(`twitch-embed-${channel}`);
-      if (!channelDiv) return;
-      channelDiv.classList.add('animate-highlight');
-      return;
-    }
-
-    if (watching.length >= 9) {
-      // do this better :]
-      alert("That's enough, dude.");
-      return;
-    }
-
-    // add player
-    setWatching([...watching, channel]);
-    setOrder([...order, channel]);
-
-    if (watching.length < 1) {
-      setActiveChat(watching[0]);
-    }
-  }
-
-  // remove channel from watching state
-  const removeWatching = useCallback(
-    (channel: string) => {
-      if (!watching.find((e) => e == channel)) return;
-
-      // remove player
-      setWatching((w) => w.filter((e) => e != channel));
-      setOrder((o) => o.filter((e) => e != channel));
-    },
-    [watching, setWatching, setOrder]
+  const { addWatching, removeWatching, reorderWatching } = useWatchingManager(
+    watching,
+    setWatching,
+    order,
+    setOrder,
+    activeChat,
+    setActiveChat
   );
-
-  // reorder channel in order state
-  function reorderWatching(channel: string, index: number, relative: boolean) {
-    console.log('watching in reorder', watching);
-    let fromOrder = order.findIndex((o) => o == channel);
-    let toOrder = relative ? fromOrder + index : index;
-    if (toOrder < 0 || toOrder > watching.length + 1) return;
-
-    // set active chat on goto first
-    if (!relative && toOrder == 0) {
-      setActiveChat(channel);
-    }
-
-    // move channel to index 0
-    let newOrder = [...order];
-    move(newOrder, fromOrder, toOrder);
-    setOrder(newOrder);
-  }
 
   // get initial followed categories
   useEffect(() => {
@@ -156,8 +100,4 @@ export function AppProvider({ children }: { children: ReactNode }) {
       {children}
     </AppContext.Provider>
   );
-}
-
-function move(order: string[], from: number, to: number) {
-  order.splice(to, 0, order.splice(from, 1)[0]);
 }
